requirements.txt
python-multipart
# YOLOv5 requirements
fastapi
uvicorn
torch>=1.7.0
torchvision>=0.8.0
Pillow>=8.0.0
